[ui] Show one asset run tag or “4 assets”, hover action for lineage
