+++ js_modules/dagster-ui/packages/ui-core/src/nav/LatestRunTag.tsx
@@ -5,7 +5,7 @@ import {Link} from 'react-router-dom';
 
 import {useQueryRefreshAtInterval, FIFTEEN_SECONDS} from '../app/QueryRefresh';
 import {RunStatus} from '../graphql/types';
-import {timingStringForStatus} from '../runs/RunDetails';
+import {timingStringForStatus} from '../runs/RunTimingDetails';
 import {RunStatusIndicator} from '../runs/RunStatusDots';
 import {DagsterTag} from '../runs/RunTag';
 import {RunTime, RUN_TIME_FRAGMENT} from '../runs/RunUtils';
