+++ js_modules/dagster-ui/packages/ui-core/src/runs/AssetKeyTagCollection.tsx
@@ -1,22 +1,13 @@
-import {
-  Box,
-  Button,
-  ButtonLink,
-  Colors,
-  Dialog,
-  DialogFooter,
-  Icon,
-  Tag,
-} from '@dagster-io/ui-components';
+import {Box, Button, Colors, Dialog, DialogFooter, Icon, Tag} from '@dagster-io/ui-components';
 import * as React from 'react';
 import {Link} from 'react-router-dom';
 
-import {displayNameForAssetKey, tokenForAssetKey} from '../asset-graph/Utils';
+import {displayNameForAssetKey} from '../asset-graph/Utils';
+import {globalAssetGraphPathForAssetsAndDescendants} from '../assets/AssetsGroupsGlobalGraphRoot';
 import {assetDetailsPathForKey} from '../assets/assetDetailsPathForKey';
 import {AssetKey} from '../assets/types';
+import {TagActionsPopover} from '../ui/TagActions';
 import {VirtualizedItemListForDialog} from '../ui/VirtualizedItemListForDialog';
-
-const MAX_ASSET_TAGS = 3;
 
 export const AssetKeyTagCollection: React.FC<{
   assetKeys: AssetKey[] | null;
@@ -29,12 +20,8 @@ export const AssetKeyTagCollection: React.FC<{
     return null;
   }
 
-  const assetCount = assetKeys.length;
-  const displayed = assetCount <= MAX_ASSET_TAGS ? assetKeys : [];
-  const hidden = assetCount - displayed.length;
-
   const showMoreDialog =
-    hidden > 0 ? (
+    assetKeys.length > 1 ? (
       <Dialog
         title={modalTitle}
         onClose={() => setShowMore(false)}
@@ -57,49 +44,72 @@ export const AssetKeyTagCollection: React.FC<{
       </Dialog>
     ) : undefined;
 
-  if (useTags) {
+  if (assetKeys.length === 1) {
+    // Outer span ensures the popover target is in the right place if the
+    // parent is a flexbox.
+    const assetKey = assetKeys[0]!;
     return (
-      <>
-        {displayed.map((assetKey, ii) => (
-          <Link to={assetDetailsPathForKey(assetKey)} key={`${tokenForAssetKey(assetKey)}-${ii}`}>
+      <span style={{lineHeight: 0}}>
+        <TagActionsPopover
+          data={{key: '', value: ''}}
+          actions={[
+            {
+              label: 'View asset',
+              to: assetDetailsPathForKey(assetKey),
+            },
+            {
+              label: 'View downstream lineage',
+              to: assetDetailsPathForKey(assetKey, {
+                view: 'lineage',
+                lineageScope: 'downstream',
+              }),
+            },
+          ]}
+        >
+          {useTags ? (
             <Tag intent="none" interactive icon="asset">
               {displayNameForAssetKey(assetKey)}
             </Tag>
-          </Link>
-        ))}
-        {hidden > 0 && (
-          <ButtonLink onClick={() => setShowMore(true)}>
-            <Tag intent="none" icon="asset">
-              {hidden} assets
-            </Tag>
-          </ButtonLink>
-        )}
-        {showMoreDialog}
-      </>
+          ) : (
+            <Box flex={{direction: 'row', gap: 8, alignItems: 'center'}}>
+              <Icon color={Colors.Gray400} name="asset" size={16} />
+              {displayNameForAssetKey(assetKey)}
+            </Box>
+          )}
+        </TagActionsPopover>
+      </span>
     );
   }
 
   return (
-    <Box flex={{direction: 'row', gap: 8, alignItems: 'center'}}>
-      <Icon color={Colors.Gray400} name="asset" size={16} />
-      <Box style={{flex: 1}} flex={{wrap: 'wrap', display: 'inline-flex'}}>
-        {displayed.map((assetKey, idx) => (
-          <Link
-            to={assetDetailsPathForKey(assetKey)}
-            key={tokenForAssetKey(assetKey)}
-            style={{marginRight: 4}}
-          >
-            {`${displayNameForAssetKey(assetKey)}${idx < displayed.length - 1 ? ',' : ''}`}
-          </Link>
-        ))}
-
-        {hidden > 0 && displayed.length > 0 ? (
-          <ButtonLink onClick={() => setShowMore(true)}>{` + ${hidden} more`}</ButtonLink>
-        ) : hidden > 0 ? (
-          <ButtonLink onClick={() => setShowMore(true)}>{`${hidden} assets`}</ButtonLink>
-        ) : undefined}
-      </Box>
+    <span style={{lineHeight: 0}}>
+      <TagActionsPopover
+        data={{key: '', value: ''}}
+        actions={[
+          {
+            label: 'View list',
+            onClick: () => setShowMore(true),
+          },
+          {
+            label: 'View downstream lineage',
+            to: globalAssetGraphPathForAssetsAndDescendants(assetKeys),
+          },
+        ]}
+      >
+        {useTags ? (
+          <Tag intent="none" icon="asset">
+            {assetKeys.length} assets
+          </Tag>
+        ) : (
+          <Box flex={{direction: 'row', gap: 8, alignItems: 'center', display: 'inline-flex'}}>
+            <Icon color={Colors.Gray400} name="asset" size={16} />
+            <Box style={{flex: 1}} flex={{wrap: 'wrap', display: 'inline-flex'}}>
+              {`${assetKeys.length} assets`}
+            </Box>
+          </Box>
+        )}
+      </TagActionsPopover>
       {showMoreDialog}
-    </Box>
+    </span>
   );
 });
