+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunFragments.tsx
@@ -3,8 +3,8 @@ import {gql} from '@apollo/client';
 import {EXECUTION_PLAN_TO_GRAPH_FRAGMENT} from '../gantt/toGraphQueryItems';
 
 import {LOGS_SCROLLING_TABLE_MESSAGE_FRAGMENT} from './LogsScrollingTable';
-import {RUN_DETAILS_FRAGMENT} from './RunDetails';
 import {RUN_METADATA_PROVIDER_MESSAGE_FRAGMENT} from './RunMetadataProvider';
+import {RUN_TIMING_FRAGMENT} from './RunTimingDetails';
 
 export const RUN_FRAGMENT = gql`
   fragment RunFragment on Run {
@@ -61,11 +61,11 @@ export const RUN_FRAGMENT = gql`
         endTime
       }
     }
-    ...RunDetailsFragment
+    ...RunTimingFragment
   }
 
   ${EXECUTION_PLAN_TO_GRAPH_FRAGMENT}
-  ${RUN_DETAILS_FRAGMENT}
+  ${RUN_TIMING_FRAGMENT}
 `;
 
 export const RUN_DAGSTER_RUN_EVENT_FRAGMENT = gql`
