+++ js_modules/dagster-ui/packages/ui-core/src/runs/types/RunTimingDetails.types.ts
@@ -2,11 +2,12 @@
 
 import * as Types from '../../graphql/types';
 
-export type RunDetailsFragment = {
+export type RunTimingFragment = {
   __typename: 'Run';
   id: string;
   startTime: number | null;
   endTime: number | null;
+  updateTime: number | null;
   status: Types.RunStatus;
   hasConcurrencyKeySlots: boolean;
 };
